finished search function and started fixing some design issues
