+++ index.html
@@ -9,28 +9,28 @@
   <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@48,400,0,0" />
 </head>
 <body>
-  <header class = "bg-dark d-flex flex-column">
-    <nav class = "navbar w-100">
-      <div class = "p-2 hidden header-profile">
+  <header class = "bg-dark w-100 m-0 p-0">
+    <nav class = "navbar w-100 container row m-auto">
+      <div class = "p-2 hidden header-profile col">
         <img src = "/media/fbprofilbilde.jpg" class = "w-10 h-10 rounded-circle profile-picture">
       </div>
-      <div class = "p2 header-login">
+      <div class = "p2 header-login col">
         <a href = "login.html" class = "text-light text-decoration-none fs-6">Log in</a>
       </div>
-      <div> 
-        <a href = "index.html" class = "text-light text-decoration-none fs-4">Home</a>
+      <div class = "col align-content-center"> 
+        <a href = "index.html" class = "text-light justify-content-center text-decoration-none fs-4 justify-self-center w-33">Home</a>
       </div>
-      <div class = "s-icon p-3">
-        <span class="material-symbols-outlined">
+      <div class = "s-icon col">
+<!--         <span class="material-symbols-outlined">
           search
-        </span>
+        </span> -->
+        <div class = "search-field w-100 ms-100">
+          <form class = "p-3 align-content-center justify-content-center d-flex search-form">
+            <input type = "text" class = "search ">
+          </form>
+        </div>
       </div>
     </nav>
-    <div class = "search-field w-100 ms-100 hidden">
-      <form class = "p-3 align-content-center justify-content-end d-flex">
-        <input type = "text" class = "search ">
-      </form>
-    </div>
   </header>
   <main>
     <h1 class = "text-center p-3">Socials Feed</h1>
